feat: [WIP] interface層の実装
